change Application

+++ Master.js
@@ -26,12 +26,15 @@ pro.init = function() {
 		for ( let i in servers ) {
 			let cfg = servers[i];
 			let logFile = 'log/'+ cfg.port +'.log';
-			fs.unlinkSync(logFile);
+			if (fs.existsSync(logFile)) {
+				fs.unlinkSync(logFile);
+			}
 			let out = fs.openSync(logFile, 'a');
 			let err = fs.openSync(logFile, 'a');
 			// this.srvDict[cfg.port] = childProcess.fork(ROOT_DIR +"Application.js", [srvId, srvType, cfg.port, cfg.clientPort]);
 			// this.srvDict[cfg.port] = childProcess.spawn("nohup", ['node', ROOT_DIR +"Application.js", srvId, srvType, cfg.port, cfg.clientPort], {stdio:[process.stdin, process.stdout, process.stderr, 'ipc']});
-			this.srvDict[cfg.port] = childProcess.spawn("node", [ROOT_DIR +"Application.js", srvId, srvType, cfg.port, cfg.clientPort], {detached: true, stdio:['ignore', out, err, 'ipc']});
+			// this.srvDict[cfg.port] = childProcess.spawn("node", [ROOT_DIR +"Application.js", srvId, srvType, cfg.port, cfg.clientPort], {detached: true, stdio:['ignore', out, err, 'ipc']});
+			this.srvDict[cfg.port] = childProcess.spawn("node", [ROOT_DIR +"Application.js", srvId, srvType, cfg.port, cfg.clientPort], {stdio:[process.stdin, process.stdout, process.stderr, 'ipc']});
 			serverMgr.add(srvId, srvType, cfg);
 			++srvId;
 			this.regMessage(this.srvDict[cfg.port], cfg.port);
